menu and auto save changes

+++ lib/views/menu_view.dart
@@ -1,10 +1,13 @@
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 import 'package:whatsapp_plugin/clipers/menu_cliper.dart';
 import 'package:whatsapp_plugin/clipers/second_menu_cliper.dart';
 import 'package:whatsapp_plugin/constants/colors.dart';
 import 'package:whatsapp_plugin/localization/app_localization.dart';
 import 'package:whatsapp_plugin/services/android_bridge.service.dart';
+import 'package:whatsapp_plugin/services/app_initializer.dart';
 import 'package:whatsapp_plugin/services/service_locator.dart';
 import 'package:whatsapp_plugin/view_models/chat_model.dart';
 
@@ -15,12 +18,13 @@ class MenuView extends StatefulWidget {
 
 class _MenuViewState extends State<MenuView> {
   AndroidBridge androidBridge = locator<AndroidBridge>();
+  AppInitializer appInitializer = locator<AppInitializer>();
+
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: Container(
-          child: Stack(
+    return Container(
+      child: Stack(
         children: <Widget>[
           ClipPath(
             child: Container(
@@ -42,8 +46,24 @@ class _MenuViewState extends State<MenuView> {
                     splashColor: Colors.lime,
                     onTap: () {
                       androidBridge.deleterAllMessages();
-                      Provider.of<ChatViewModel>(context, listen: false)
-                          .deleteAllMessage();
+                      showDialog(
+                        context: context,
+                        barrierDismissible: false,
+                        builder: (_) => AlertDialog(
+                          title: Text('Do you want to delete all message?'),
+                          actions: <Widget>[
+                            FlatButton(child: Text('No'),onPressed: () {
+                              Navigator.pop(context);
+                            },),
+                            FlatButton(child: Text('Yes'),onPressed: (){
+                              Provider.of<ChatViewModel>(context, listen: false)
+                                  .deleteAllMessage();
+                              Navigator.pop(context);
+                            },),
+
+                          ],
+                        ),
+                      );
                       print('deleye msg');
                     },
                     child: ListTile(
@@ -53,7 +73,8 @@ class _MenuViewState extends State<MenuView> {
                         color: Colors.white,
                       ),
                       title: Text(
-                        AppLocalizations.of(context).localizedValues['delete_message'],
+                        AppLocalizations.of(context)
+                            .localizedValues['delete_message'],
                         style: TextStyle(color: Colors.white),
                       ),
                     ),
@@ -63,7 +84,7 @@ class _MenuViewState extends State<MenuView> {
                     child: InkWell(
                       splashColor: Colors.tealAccent,
                       onTap: () {
-                        print("tapped");
+                        Navigator.pushNamed(context, '/faq');
                       },
                       child: Container(
                         child: ListTile(
@@ -82,33 +103,39 @@ class _MenuViewState extends State<MenuView> {
                   ),
                   Material(
                     color: Colors.transparent,
-                    child: InkWell(
-                      splashColor: Colors.tealAccent,
-                      onTap: () {
-                        print("tapped");
-                      },
-                      child: Container(
-                        child: ListTile(
-                          enabled: true,
-                          leading: const Icon(
-                            Icons.visibility,
-                            color: Colors.white,
-                          ),
-                          title: Text(
-                            AppLocalizations.of(context).localizedValues['privacy_policy'],
-                            style: TextStyle(color: Colors.white),
-                          ),
+                    child: Container(
+                      child: ListTile(
+                        enabled: true,
+                        leading: const Icon(
+                          Icons.save_alt,
+                          color: Colors.white,
                         ),
+                        title: Text(
+                          AppLocalizations.of(context).localizedValues['auto_save'],
+                          style: TextStyle(color: Colors.white),
+                        ),
+                        trailing: Switch(value: appInitializer.autoSaveStatus, onChanged: (bool value) {
+                          setState(() {
+                            appInitializer.autoSaveStatus = value;
+                            if (value == false) {
+                              print('Service Stoped');
+                              androidBridge.stopListenToStatusGen();
+                            } else {
+                              print('Service Started');
+                              androidBridge.startListenToStatusGen();
+                            }
+                          });
+                        }),
                       ),
                     ),
-                  )
+                  ),
                 ],
               ),
             ),
             clipper: SecondMenuCliper(),
           )
         ],
-      )),
+      ),
     );
   }
 }
